requirements.txt
opencc-python-reimplemented
pymongo
stanfordnlp
ansi
pandas
numpy